+++ crates/ruma-identifiers/src/room_or_room_alias_id.rs
@@ -9,33 +9,33 @@ use crate::{server_name::ServerName, RoomAliasId, RoomId};
 
 /// A Matrix room ID or a Matrix room alias ID.
 ///
-/// `RoomIdOrAliasId` is useful for APIs that accept either kind of room identifier. It is converted
+/// `RoomOrAliasId` is useful for APIs that accept either kind of room identifier. It is converted
 /// from a string slice, and can be converted back into a string as needed. When converted from a
 /// string slice, the variant is determined by the leading sigil character.
 ///
 /// ```
 /// # use std::convert::TryFrom;
-/// # use ruma_identifiers::RoomIdOrAliasId;
+/// # use ruma_identifiers::RoomOrAliasId;
 /// assert_eq!(
-///     <&RoomIdOrAliasId>::try_from("#ruma:example.com").unwrap(),
+///     <&RoomOrAliasId>::try_from("#ruma:example.com").unwrap(),
 ///     "#ruma:example.com"
 /// );
 ///
 /// assert_eq!(
-///     <&RoomIdOrAliasId>::try_from("!n8f893n9:example.com").unwrap(),
+///     <&RoomOrAliasId>::try_from("!n8f893n9:example.com").unwrap(),
 ///     "!n8f893n9:example.com"
 /// );
 /// ```
 #[repr(transparent)]
 #[derive(PartialEq, Eq, PartialOrd, Ord, Hash)]
-pub struct RoomIdOrAliasId(str);
+pub struct RoomOrAliasId(str);
 
 opaque_identifier_validated!(
-    RoomIdOrAliasId,
+    RoomOrAliasId,
     ruma_identifiers_validation::room_id_or_alias_id::validate
 );
 
-impl RoomIdOrAliasId {
+impl RoomOrAliasId {
     /// Returns the local part (everything after the `!` or `#` and before the first colon).
     pub fn localpart(&self) -> &str {
         &self.as_str()[1..self.colon_idx()]
@@ -56,7 +56,7 @@ impl RoomIdOrAliasId {
         self.variant() == Variant::RoomAliasId
     }
 
-    /// Turn this `RoomIdOrAliasId` into `Either<RoomId, RoomAliasId>`
+    /// Turn this `RoomOrAliasId` into `Either<RoomId, RoomAliasId>`
     #[cfg(feature = "either")]
     pub fn into_either(self: Box<Self>) -> either::Either<Box<RoomId>, Box<RoomAliasId>> {
         let variant = self.variant();
@@ -87,22 +87,22 @@ enum Variant {
     RoomAliasId,
 }
 
-impl From<Box<RoomId>> for Box<RoomIdOrAliasId> {
+impl From<Box<RoomId>> for Box<RoomOrAliasId> {
     fn from(room_id: Box<RoomId>) -> Self {
         Self::try_from(room_id.as_str()).unwrap()
     }
 }
 
-impl From<Box<RoomAliasId>> for Box<RoomIdOrAliasId> {
+impl From<Box<RoomAliasId>> for Box<RoomOrAliasId> {
     fn from(room_alias_id: Box<RoomAliasId>) -> Self {
         Self::try_from(room_alias_id.as_str()).unwrap()
     }
 }
 
-impl TryFrom<Box<RoomIdOrAliasId>> for Box<RoomId> {
+impl TryFrom<Box<RoomOrAliasId>> for Box<RoomId> {
     type Error = Box<RoomAliasId>;
 
-    fn try_from(id: Box<RoomIdOrAliasId>) -> Result<Box<RoomId>, Box<RoomAliasId>> {
+    fn try_from(id: Box<RoomOrAliasId>) -> Result<Box<RoomId>, Box<RoomAliasId>> {
         match id.variant() {
             Variant::RoomId => Ok(id.as_str().try_into().unwrap()),
             Variant::RoomAliasId => Err(id.as_str().try_into().unwrap()),
@@ -110,10 +110,10 @@ impl TryFrom<Box<RoomIdOrAliasId>> for Box<RoomId> {
     }
 }
 
-impl TryFrom<Box<RoomIdOrAliasId>> for Box<RoomAliasId> {
+impl TryFrom<Box<RoomOrAliasId>> for Box<RoomAliasId> {
     type Error = Box<RoomId>;
 
-    fn try_from(id: Box<RoomIdOrAliasId>) -> Result<Box<RoomAliasId>, Box<RoomId>> {
+    fn try_from(id: Box<RoomOrAliasId>) -> Result<Box<RoomAliasId>, Box<RoomId>> {
         match id.variant() {
             Variant::RoomAliasId => Ok(id.as_str().try_into().unwrap()),
             Variant::RoomId => Err(id.as_str().try_into().unwrap()),
@@ -125,13 +125,13 @@ impl TryFrom<Box<RoomIdOrAliasId>> for Box<RoomAliasId> {
 mod tests {
     use std::convert::TryFrom;
 
-    use super::RoomIdOrAliasId;
+    use super::RoomOrAliasId;
     use crate::Error;
 
     #[test]
     fn valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
-            <&RoomIdOrAliasId>::try_from("#ruma:example.com")
+            <&RoomOrAliasId>::try_from("#ruma:example.com")
                 .expect("Failed to create RoomAliasId.")
                 .as_ref(),
             "#ruma:example.com"
@@ -141,7 +141,7 @@ mod tests {
     #[test]
     fn valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
-            <&RoomIdOrAliasId>::try_from("!29fhd83h92h0:example.com")
+            <&RoomOrAliasId>::try_from("!29fhd83h92h0:example.com")
                 .expect("Failed to create RoomId.")
                 .as_ref(),
             "!29fhd83h92h0:example.com"
@@ -151,7 +151,7 @@ mod tests {
     #[test]
     fn missing_sigil_for_room_id_or_alias_id() {
         assert_eq!(
-            <&RoomIdOrAliasId>::try_from("ruma:example.com").unwrap_err(),
+            <&RoomOrAliasId>::try_from("ruma:example.com").unwrap_err(),
             Error::MissingLeadingSigil
         );
     }
@@ -161,7 +161,7 @@ mod tests {
     fn serialize_valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
             serde_json::to_string(
-                <&RoomIdOrAliasId>::try_from("#ruma:example.com")
+                <&RoomOrAliasId>::try_from("#ruma:example.com")
                     .expect("Failed to create RoomAliasId.")
             )
             .expect("Failed to convert RoomAliasId to JSON."),
@@ -174,7 +174,7 @@ mod tests {
     fn serialize_valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
             serde_json::to_string(
-                <&RoomIdOrAliasId>::try_from("!29fhd83h92h0:example.com")
+                <&RoomOrAliasId>::try_from("!29fhd83h92h0:example.com")
                     .expect("Failed to create RoomId.")
             )
             .expect("Failed to convert RoomId to JSON."),
@@ -186,10 +186,9 @@ mod tests {
     #[test]
     fn deserialize_valid_room_id_or_alias_id_with_a_room_alias_id() {
         assert_eq!(
-            serde_json::from_str::<Box<RoomIdOrAliasId>>(r##""#ruma:example.com""##)
+            serde_json::from_str::<Box<RoomOrAliasId>>(r##""#ruma:example.com""##)
                 .expect("Failed to convert JSON to RoomAliasId"),
-            <&RoomIdOrAliasId>::try_from("#ruma:example.com")
-                .expect("Failed to create RoomAliasId.")
+            <&RoomOrAliasId>::try_from("#ruma:example.com").expect("Failed to create RoomAliasId.")
         );
     }
 
@@ -197,9 +196,9 @@ mod tests {
     #[test]
     fn deserialize_valid_room_id_or_alias_id_with_a_room_id() {
         assert_eq!(
-            serde_json::from_str::<Box<RoomIdOrAliasId>>(r##""!29fhd83h92h0:example.com""##)
+            serde_json::from_str::<Box<RoomOrAliasId>>(r##""!29fhd83h92h0:example.com""##)
                 .expect("Failed to convert JSON to RoomId"),
-            <&RoomIdOrAliasId>::try_from("!29fhd83h92h0:example.com")
+            <&RoomOrAliasId>::try_from("!29fhd83h92h0:example.com")
                 .expect("Failed to create RoomAliasId.")
         );
     }